refactor(studio): split viewer/state and ui_panels/selected into focused modules
Both files had grown past 1600 lines and mixed unrelated concerns. Splits
preserve external visibility (re-exported via the original module path)
and behaviour (cargo test -p mogen-studio: 215/215 passing).

- viewer/state.rs (1684 lines) → viewer/state/{capture,gizmo,selection,path,mod}.rs:
  capture types + PNG encode pool, gizmo drag pipeline + snap helpers,
  multi-selection bookkeeping + redirect_pick, and stable path lookups.
- app/ui_panels/selected.rs (1993 lines) →
  app/ui_panels/selected/{deform_rows,geom_params,use_wrap,mod}.rs:
  the ui_selected impl stays in mod.rs; row renderers, primitive-param
  switch, and the use-wrap/kind-rewrite helpers move to siblings.
