+++ crates/mogen-studio/src/app/ui_panels/selected/mod.rs
@@ -6,6 +6,16 @@ use crate::app::types::UndoKey;
 use crate::app::util::format_inspector_scalar;
 use crate::app::MogenStudioApp;
 
+mod deform_rows;
+mod geom_params;
+mod use_wrap;
+
+use deform_rows::{
+    deform_angle_row, deform_faceted_row, deform_seed_row, deform_taper_row, deform_unit_row,
+};
+use geom_params::geom_params_for_kind;
+use use_wrap::{resolve_use_wrap_target, rewrite_node_kind};
+
 impl MogenStudioApp {
     /// Inspector for the currently-selected scene node. Shown above the
     /// scene summary; collapses to a friendly hint when nothing's selected
@@ -1415,579 +1425,4 @@ impl MogenStudioApp {
             }
         });
     }
-}
-
-/// Resolve the active-source span of the `use "..." (...)` call that brought
-/// `node` into the scene, plus a sensible group name to wrap it under. Used
-/// by the "Wrap `use` in a group" affordance shown next to the
-/// imported-via-use warning.
-///
-/// Two cases:
-///   - **Imported file (`origin = Some`)**: the file's stem (e.g.
-///     `humanoid_full.mog` → `humanoid_full`) matches the use's name. The
-///     first `use "<stem>"` AST node in the active source wins.
-///   - **Local module (`origin = None`)**: the use call lives in the body of
-///     the closest user-authored ancestor (the first ancestor in the chain
-///     with `use_id != node.use_id` that has a span). We pick the first
-///     `use` AST child within that ancestor's span. With multiple use calls
-///     the first match isn't always *the* call that minted this node, but
-///     undo recovers and the warning only fires when there's no wrapper, so
-///     ambiguity is uncommon.
-///
-/// Returns `None` when the source no longer parses or no candidate `use`
-/// declaration is found.
-fn resolve_use_wrap_target(
-    scene: &mogen_core::SceneGraph,
-    sel: mogen_core::NodeId,
-    node: &mogen_core::SceneNode,
-    source: &str,
-) -> Option<(mogen_core::Span, String)> {
-    if let Some(stem) = node
-        .origin
-        .as_deref()
-        .and_then(|p| p.file_stem())
-        .and_then(|s| s.to_str())
-    {
-        let span = crate::app::util::find_use_source_span(source, stem)?;
-        return Some((span, stem.to_string()));
-    }
-
-    // Local-module case: walk up to the closest ancestor whose `use_id`
-    // differs from the selected node's. That ancestor's source span (in the
-    // active file) bounds where the originating `use` call lives.
-    let mut cur = node.parent;
-    let mut ancestor_span: Option<mogen_core::Span> = None;
-    while let Some(pid) = cur {
-        let parent = scene.nodes.get(pid.0 as usize)?;
-        if parent.use_id != node.use_id {
-            ancestor_span = parent.source_span;
-            break;
-        }
-        cur = parent.parent;
-    }
-    let ancestor_span = ancestor_span?;
-    let ast = mogen_dsl::parse(source).ok()?;
-    let _ = sel;
-
-    fn find_first_use_within(
-        nodes: &[mogen_dsl::ast::Node],
-        bounds: mogen_core::Span,
-    ) -> Option<&mogen_dsl::ast::Node> {
-        for n in nodes {
-            if n.span.start < bounds.start || n.span.end > bounds.end {
-                continue;
-            }
-            if n.kind == "use" {
-                return Some(n);
-            }
-            if let Some(child) = find_first_use_within(&n.children, bounds) {
-                return Some(child);
-            }
-        }
-        None
-    }
-    let use_node = find_first_use_within(&ast, ancestor_span)?;
-    let name = use_node.name.clone()?;
-    Some((use_node.span, name))
-}
-
-/// Render a Deform-row for a 0..=1 unit modifier (`noise`, `jitter`, `droop`).
-/// `default` is the value used when the user clicks "+ add" on an absent attr.
-#[allow(clippy::too_many_arguments)]
-fn deform_unit_row(
-    ui: &mut egui::Ui,
-    label: &str,
-    attr: &'static str,
-    current: Option<f32>,
-    default: f32,
-    node_id: mogen_core::NodeId,
-    edits: &mut Vec<crate::viewer::PendingEdit>,
-    wants_remove: &mut Vec<&'static str>,
-) {
-    deform_scalar_row(
-        ui, label, attr, current, default,
-        0.0..=1.0, 0.01, "", node_id, edits, wants_remove,
-    );
-}
-
-/// Render a Deform-row for an angle modifier (`bend_x`/`y`/`z`, `twist_y`),
-/// authored in degrees. The lowering pass runs `.to_radians()` before applying.
-#[allow(clippy::too_many_arguments)]
-fn deform_angle_row(
-    ui: &mut egui::Ui,
-    label: &str,
-    attr: &'static str,
-    current: Option<f32>,
-    default: f32,
-    node_id: mogen_core::NodeId,
-    edits: &mut Vec<crate::viewer::PendingEdit>,
-    wants_remove: &mut Vec<&'static str>,
-) {
-    deform_scalar_row(
-        ui, label, attr, current, default,
-        -180.0..=180.0, 0.5, "°", node_id, edits, wants_remove,
-    );
-}
-
-/// Render the `taper` row. Different range/default from the unit row because
-/// taper's neutral is 1.0 (no scale change) and authors flare past 1.0 for
-/// trumpet shapes — clamping to [0, 1] would silently drop that case.
-fn deform_taper_row(
-    ui: &mut egui::Ui,
-    current: Option<f32>,
-    node_id: mogen_core::NodeId,
-    edits: &mut Vec<crate::viewer::PendingEdit>,
-    wants_remove: &mut Vec<&'static str>,
-) {
-    deform_scalar_row(
-        ui, "Taper", "taper", current, 0.5,
-        0.0..=4.0, 0.02, "", node_id, edits, wants_remove,
-    );
-}
-
-#[allow(clippy::too_many_arguments)]
-fn deform_scalar_row(
-    ui: &mut egui::Ui,
-    label: &str,
-    attr: &'static str,
-    current: Option<f32>,
-    default: f32,
-    range: std::ops::RangeInclusive<f32>,
-    speed: f32,
-    suffix: &str,
-    node_id: mogen_core::NodeId,
-    edits: &mut Vec<crate::viewer::PendingEdit>,
-    wants_remove: &mut Vec<&'static str>,
-) {
-    use crate::app::util::format_inspector_scalar;
-    use crate::viewer::PendingEdit;
-    ui.label(label);
-    ui.horizontal(|ui| match current {
-        Some(initial) => {
-            let mut v = initial;
-            let resp = ui.add(
-                egui::DragValue::new(&mut v)
-                    .speed(speed)
-                    .suffix(suffix)
-                    .range(range),
-            );
-            if resp.changed() {
-                edits.push(PendingEdit::SetAttrCanonical {
-                    node: node_id,
-                    attr: attr.into(),
-                    value: format_inspector_scalar(v),
-                    delete: Vec::new(),
-                });
-            }
-            if ui
-                .small_button("✕")
-                .on_hover_text("Remove modifier")
-                .clicked()
-            {
-                wants_remove.push(attr);
-            }
-        }
-        None => {
-            ui.label(egui::RichText::new("(none)").italics().weak());
-            if ui.small_button("+ add").clicked() {
-                edits.push(PendingEdit::SetAttrCanonical {
-                    node: node_id,
-                    attr: attr.into(),
-                    value: format_inspector_scalar(default),
-                    delete: Vec::new(),
-                });
-            }
-        }
-    });
-    ui.end_row();
-}
-
-/// Render the `faceted` row as a checkbox: present-and-non-zero is on, absent
-/// or `0` is off. Toggling on writes `faceted=1`; toggling off removes the
-/// attr entirely so it doesn't sit at `0` polluting the source.
-fn deform_faceted_row(
-    ui: &mut egui::Ui,
-    current: Option<f32>,
-    node_id: mogen_core::NodeId,
-    edits: &mut Vec<crate::viewer::PendingEdit>,
-    wants_remove: &mut Vec<&'static str>,
-) {
-    use crate::viewer::PendingEdit;
-    ui.label("Faceted");
-    ui.horizontal(|ui| {
-        let on = current.map(|v| v != 0.0).unwrap_or(false);
-        let mut new_on = on;
-        ui.checkbox(&mut new_on, "")
-            .on_hover_text("Hard-edge per-triangle normals (low-poly look).");
-        if new_on != on {
-            if new_on {
-                edits.push(PendingEdit::SetAttrCanonical {
-                    node: node_id,
-                    attr: "faceted".into(),
-                    value: "1".into(),
-                    delete: Vec::new(),
-                });
-            } else if current.is_some() {
-                wants_remove.push("faceted");
-            }
-        }
-    });
-    ui.end_row();
-}
-
-/// Render the `seed` row. Drives the random stream for `noise` / `jitter`;
-/// authors typically tweak it to roll a different stochastic result while
-/// keeping the modifier amounts the same.
-fn deform_seed_row(
-    ui: &mut egui::Ui,
-    current: Option<u32>,
-    node_id: mogen_core::NodeId,
-    edits: &mut Vec<crate::viewer::PendingEdit>,
-    wants_remove: &mut Vec<&'static str>,
-) {
-    use crate::viewer::PendingEdit;
-    ui.label("Seed");
-    ui.horizontal(|ui| match current {
-        Some(initial) => {
-            let mut v = initial;
-            let resp = ui.add(
-                egui::DragValue::new(&mut v)
-                    .speed(1.0)
-                    .range(0..=u32::MAX),
-            );
-            if resp.changed() {
-                edits.push(PendingEdit::SetAttrCanonical {
-                    node: node_id,
-                    attr: "seed".into(),
-                    value: v.to_string(),
-                    delete: Vec::new(),
-                });
-            }
-            if ui
-                .small_button("✕")
-                .on_hover_text("Remove seed (defaults to 1)")
-                .clicked()
-            {
-                wants_remove.push("seed");
-            }
-        }
-        None => {
-            ui.label(egui::RichText::new("(none)").italics().weak());
-            if ui.small_button("+ add").clicked() {
-                edits.push(PendingEdit::SetAttrCanonical {
-                    node: node_id,
-                    attr: "seed".into(),
-                    value: "1".into(),
-                    delete: Vec::new(),
-                });
-            }
-        }
-    });
-    ui.end_row();
-}
-
-/// Render scalar geometry-parameter rows for `kind`. Returns `true` when at
-/// least one editable row was added, so the caller can paint a "(no editable
-/// params)" hint when the kind has nothing to show. Each numeric attr falls
-/// back to its primitive lowering default — see `mogen_dsl::lower::primitive`.
-/// List-shaped attrs (`points`, `profile`, `path`, `holes`) are intentionally
-/// skipped — they need a richer editor than a sidebar grid.
-///
-/// Out entries are pre-formatted `(attr_name, value_literal)` pairs so the
-/// helper can emit list-shaped writes (`size=[1, 2, 3]`) without the caller
-/// having to know which attrs are scalar vs. vector.
-#[allow(clippy::too_many_arguments)]
-pub(in crate::app) fn geom_params_for_kind(
-    ui: &mut egui::Ui,
-    kind: &str,
-    src: &str,
-    span: mogen_core::Span,
-    out: &mut Vec<(&'static str, String)>,
-) -> bool {
-    use crate::edit::get_attr;
-    fn read(src: &str, span: mogen_core::Span, attr: &str) -> Option<f32> {
-        get_attr(src, span, attr).and_then(|s| s.parse::<f32>().ok())
-    }
-    fn read_vec3(src: &str, span: mogen_core::Span, attr: &str) -> Option<[f32; 3]> {
-        let raw = get_attr(src, span, attr)?;
-        let trimmed = raw.trim().trim_start_matches('[').trim_end_matches(']');
-        let parts: Vec<f32> = trimmed
-            .split(',')
-            .filter_map(|s| s.trim().parse::<f32>().ok())
-            .collect();
-        match parts.len() {
-            1 => Some([parts[0], parts[0], parts[0]]),
-            3 => Some([parts[0], parts[1], parts[2]]),
-            _ => None,
-        }
-    }
-    fn read_vec2(src: &str, span: mogen_core::Span, attr: &str) -> Option<[f32; 2]> {
-        let raw = get_attr(src, span, attr)?;
-        let trimmed = raw.trim().trim_start_matches('[').trim_end_matches(']');
-        let parts: Vec<f32> = trimmed
-            .split(',')
-            .filter_map(|s| s.trim().parse::<f32>().ok())
-            .collect();
-        match parts.len() {
-            1 => Some([parts[0], parts[0]]),
-            2 => Some([parts[0], parts[1]]),
-            _ => None,
-        }
-    }
-    fn scalar_row(
-        ui: &mut egui::Ui,
-        label: &str,
-        attr: &'static str,
-        current: f32,
-        speed: f32,
-        out: &mut Vec<(&'static str, String)>,
-    ) {
-        use crate::app::util::format_inspector_scalar;
-        ui.label(label);
-        let mut v = current;
-        if ui.add(egui::DragValue::new(&mut v).speed(speed)).changed() {
-            out.push((attr, format_inspector_scalar(v)));
-        }
-        ui.end_row();
-    }
-    fn int_row(
-        ui: &mut egui::Ui,
-        label: &str,
-        attr: &'static str,
-        current: i32,
-        min: i32,
-        max: i32,
-        out: &mut Vec<(&'static str, String)>,
-    ) {
-        ui.label(label);
-        let mut v = current;
-        if ui
-            .add(egui::DragValue::new(&mut v).speed(0.1).range(min..=max))
-            .changed()
-        {
-            out.push((attr, v.to_string()));
-        }
-        ui.end_row();
-    }
-    use crate::app::util::format_inspector_scalar;
-    let mut shown = false;
-    match kind {
-        "box" | "slab" | "post" | "panel" | "wedge" | "ellipsoid" | "rounded_box"
-        | "chamfered_box" | "inset_box" | "wall" | "prism" | "superellipsoid" => {
-            let s = read_vec3(src, span, "size").unwrap_or([1.0, 1.0, 1.0]);
-            ui.label("Size");
-            ui.horizontal(|ui| {
-                let mut sx = s[0];
-                let mut sy = s[1];
-                let mut sz = s[2];
-                let mut emit = false;
-                if ui.add(egui::DragValue::new(&mut sx).speed(0.02)).changed() { emit = true; }
-                if ui.add(egui::DragValue::new(&mut sy).speed(0.02)).changed() { emit = true; }
-                if ui.add(egui::DragValue::new(&mut sz).speed(0.02)).changed() { emit = true; }
-                if emit {
-                    out.push((
-                        "size",
-                        format!(
-                            "[{}, {}, {}]",
-                            format_inspector_scalar(sx),
-                            format_inspector_scalar(sy),
-                            format_inspector_scalar(sz),
-                        ),
-                    ));
-                }
-            });
-            ui.end_row();
-            shown = true;
-            // Kind-specific extras
-            match kind {
-                "rounded_box" | "chamfered_box" => {
-                    scalar_row(ui, "Radius", "radius",
-                        read(src, span, "radius").unwrap_or(0.1), 0.005, out);
-                    if kind == "rounded_box" {
-                        let segs = read(src, span, "segments").unwrap_or(4.0) as i32;
-                        int_row(ui, "Segments", "segments", segs, 1, 32, out);
-                    }
-                }
-                "inset_box" => {
-                    scalar_row(ui, "Amount", "amount",
-                        read(src, span, "amount").unwrap_or(0.1), 0.005, out);
-                    scalar_row(ui, "Depth", "depth",
-                        read(src, span, "depth").unwrap_or(0.05), 0.005, out);
-                }
-                "superellipsoid" => {
-                    scalar_row(ui, "Equator", "ew",
-                        read(src, span, "ew").unwrap_or(1.0), 0.05, out);
-                    scalar_row(ui, "Meridian", "ns",
-                        read(src, span, "ns").unwrap_or(1.0), 0.05, out);
-                    let r = read(src, span, "rings").unwrap_or(16.0) as i32;
-                    let s = read(src, span, "segments").unwrap_or(24.0) as i32;
-                    int_row(ui, "Rings", "rings", r, 2, 256, out);
-                    int_row(ui, "Segments", "segments", s, 3, 256, out);
-                }
-                "ellipsoid" => {
-                    let r = read(src, span, "rings").unwrap_or(16.0) as i32;
-                    let s = read(src, span, "segments").unwrap_or(24.0) as i32;
-                    int_row(ui, "Rings", "rings", r, 2, 256, out);
-                    int_row(ui, "Segments", "segments", s, 3, 256, out);
-                }
-                _ => {}
-            }
-        }
-        "plane" | "quad" | "decal" | "leaf_card" | "curved_plane" => {
-            // plane uses [x,z], quad/decal/leaf_card uses [x,y], curved_plane uses [x,z].
-            let s = read_vec2(src, span, "size").unwrap_or([1.0, 1.0]);
-            ui.label("Size");
-            ui.horizontal(|ui| {
-                let mut su = s[0];
-                let mut sv = s[1];
-                let mut emit = false;
-                if ui.add(egui::DragValue::new(&mut su).speed(0.05)).changed() { emit = true; }
-                if ui.add(egui::DragValue::new(&mut sv).speed(0.05)).changed() { emit = true; }
-                if emit {
-                    out.push((
-                        "size",
-                        format!(
-                            "[{}, {}]",
-                            format_inspector_scalar(su),
-                            format_inspector_scalar(sv),
-                        ),
-                    ));
-                }
-            });
-            ui.end_row();
-            shown = true;
-            if kind == "decal" {
-                scalar_row(ui, "Offset", "offset",
-                    read(src, span, "offset").unwrap_or(0.001), 0.0005, out);
-            }
-        }
-        "cylinder" | "cone" | "half_cylinder" => {
-            scalar_row(ui, "Radius", "radius",
-                read(src, span, "radius").unwrap_or(0.5), 0.02, out);
-            scalar_row(ui, "Height", "height",
-                read(src, span, "height").unwrap_or(1.0), 0.02, out);
-            let s = read(src, span, "segments").unwrap_or(24.0) as i32;
-            int_row(ui, "Segments", "segments", s, 3, 256, out);
-            shown = true;
-        }
-        "sphere" | "hemisphere" => {
-            scalar_row(ui, "Radius", "radius",
-                read(src, span, "radius").unwrap_or(0.5), 0.02, out);
-            let r = read(src, span, "rings").unwrap_or(if kind == "hemisphere" { 8.0 } else { 16.0 }) as i32;
-            let s = read(src, span, "segments").unwrap_or(24.0) as i32;
-            int_row(ui, "Rings", "rings", r, 2, 256, out);
-            int_row(ui, "Segments", "segments", s, 3, 256, out);
-            shown = true;
-        }
-        "icosphere" => {
-            scalar_row(ui, "Radius", "radius",
-                read(src, span, "radius").unwrap_or(0.5), 0.02, out);
-            let s = read(src, span, "subdivisions").unwrap_or(2.0) as i32;
-            int_row(ui, "Subdivisions", "subdivisions", s, 0, 6, out);
-            shown = true;
-        }
-        "capsule" => {
-            scalar_row(ui, "Radius", "radius",
-                read(src, span, "radius").unwrap_or(0.5), 0.02, out);
-            scalar_row(ui, "Height", "height",
-                read(src, span, "height").unwrap_or(1.0), 0.02, out);
-            let r = read(src, span, "rings").unwrap_or(8.0) as i32;
-            let s = read(src, span, "segments").unwrap_or(24.0) as i32;
-            int_row(ui, "Rings", "rings", r, 2, 64, out);
-            int_row(ui, "Segments", "segments", s, 3, 256, out);
-            shown = true;
-        }
-        "torus" => {
-            scalar_row(ui, "Major", "major",
-                read(src, span, "major").unwrap_or(0.5), 0.02, out);
-            scalar_row(ui, "Minor", "minor",
-                read(src, span, "minor").unwrap_or(0.15), 0.02, out);
-            let mj = read(src, span, "major_segments").unwrap_or(24.0) as i32;
-            let mn = read(src, span, "minor_segments").unwrap_or(12.0) as i32;
-            int_row(ui, "Major segs", "major_segments", mj, 3, 256, out);
-            int_row(ui, "Minor segs", "minor_segments", mn, 3, 256, out);
-            shown = true;
-        }
-        "torus_arc" => {
-            scalar_row(ui, "Major", "major",
-                read(src, span, "major").unwrap_or(0.5), 0.02, out);
-            scalar_row(ui, "Minor", "minor",
-                read(src, span, "minor").unwrap_or(0.15), 0.02, out);
-            scalar_row(ui, "Arc°", "arc",
-                read(src, span, "arc").unwrap_or(90.0), 1.0, out);
-            shown = true;
-        }
-        "pyramid" => {
-            scalar_row(ui, "Radius", "radius",
-                read(src, span, "radius").unwrap_or(0.5), 0.02, out);
-            scalar_row(ui, "Height", "height",
-                read(src, span, "height").unwrap_or(1.0), 0.02, out);
-            let s = read(src, span, "sides").unwrap_or(4.0) as i32;
-            int_row(ui, "Sides", "sides", s, 3, 64, out);
-            shown = true;
-        }
-        "disc" => {
-            scalar_row(ui, "Radius", "radius",
-                read(src, span, "radius").unwrap_or(0.5), 0.02, out);
-            let s = read(src, span, "segments").unwrap_or(24.0) as i32;
-            int_row(ui, "Segments", "segments", s, 3, 256, out);
-            shown = true;
-        }
-        "tube" => {
-            scalar_row(ui, "Outer", "outer",
-                read(src, span, "outer").unwrap_or(0.5), 0.02, out);
-            scalar_row(ui, "Inner", "inner",
-                read(src, span, "inner").unwrap_or(0.3), 0.02, out);
-            scalar_row(ui, "Height", "height",
-                read(src, span, "height").unwrap_or(1.0), 0.02, out);
-            let s = read(src, span, "segments").unwrap_or(24.0) as i32;
-            int_row(ui, "Segments", "segments", s, 3, 256, out);
-            shown = true;
-        }
-        "coil" => {
-            scalar_row(ui, "Radius", "radius",
-                read(src, span, "radius").unwrap_or(0.5), 0.02, out);
-            scalar_row(ui, "Height", "height",
-                read(src, span, "height").unwrap_or(1.0), 0.02, out);
-            scalar_row(ui, "Turns", "turns",
-                read(src, span, "turns").unwrap_or(3.0), 0.1, out);
-            scalar_row(ui, "Profile r", "profile_radius",
-                read(src, span, "profile_radius").unwrap_or(0.05), 0.005, out);
-            shown = true;
-        }
-        "frustum" => {
-            scalar_row(ui, "Height", "height",
-                read(src, span, "height").unwrap_or(1.0), 0.02, out);
-            shown = true;
-        }
-        _ => {}
-    }
-    shown
-}
-
-/// Rewrite the keyword (kind identifier) at the very start of the node
-/// covered by `span`. Used by the CSG op switch in the inspector — the
-/// keyword sits before the optional `"name" (...)` header, and existing
-/// edit helpers don't expose a kind rewrite. Bytes-only; no AST round-trip.
-fn rewrite_node_kind(src: &str, span: mogen_core::Span, new_kind: &str) -> String {
-    let bytes = src.as_bytes();
-    let start = span.start.min(src.len());
-    // Skip leading whitespace inside the span — pest spans typically begin
-    // exactly at the kind keyword, but be defensive.
-    let mut i = start;
-    while i < bytes.len() && bytes[i].is_ascii_whitespace() {
-        i += 1;
-    }
-    let kind_start = i;
-    while i < bytes.len() && (bytes[i].is_ascii_alphanumeric() || bytes[i] == b'_') {
-        i += 1;
-    }
-    if i == kind_start {
-        return src.to_string();
-    }
-    let mut out = String::with_capacity(src.len() + new_kind.len());
-    out.push_str(&src[..kind_start]);
-    out.push_str(new_kind);
-    out.push_str(&src[i..]);
-    out
 }
